+++ src/main/java/com/createver/server/domain/image/service/avatar/ImageAvatarProcessingService.java
@@ -1,4 +1,4 @@
-package com.createver.server.domain.image.service;
+package com.createver.server.domain.image.service.avatar;
 
 import com.createver.server.domain.image.dto.response.ImageAvatarWebhookResponse;
 import com.createver.server.domain.image.entity.ImageAvatar;
