+++ src/main/java/com/createver/server/domain/image/service/gallery/GalleryService.java
@@ -1,4 +1,4 @@
-package com.createver.server.domain.image.service;
+package com.createver.server.domain.image.service.gallery;
 
 import com.createver.server.domain.image.dto.GalleryDto;
 import com.createver.server.domain.image.repository.gallery.GalleryRepository;
