+++ src/main/java/com/createver/server/domain/image/service/gallery/ImageGenerationService.java
@@ -1,4 +1,4 @@
-package com.createver.server.domain.image.service;
+package com.createver.server.domain.image.service.gallery;
 
 import com.amazonaws.services.s3.model.AmazonS3Exception;
 import com.createver.server.domain.image.dto.request.ImageGenerationRequest;
@@ -9,6 +9,7 @@ import com.createver.server.domain.image.dto.response.ImageGenerationResponse;
 import com.createver.server.domain.image.entity.Gallery;
 import com.createver.server.domain.image.entity.ImageTag;
 import com.createver.server.domain.image.repository.gallery.GalleryRepository;
+import com.createver.server.domain.image.service.tag.ImageTagService;
 import com.createver.server.domain.member.entity.Member;
 import com.createver.server.domain.member.repository.MemberRepository;
 import com.createver.server.global.config.OpenAiConfig;
