+++ src/test/java/com/createver/server/domain/image/controller/ImageAvatarControllerTest.java
@@ -2,9 +2,9 @@ package com.createver.server.domain.image.controller;
 
 import com.createver.server.domain.image.dto.request.AvatarPromptRequest;
 import com.createver.server.domain.image.dto.response.ImageAvatarWebhookResponse;
-import com.createver.server.domain.image.service.ImageAvatarProcessingService;
-import com.createver.server.domain.image.service.ImageAvatarService;
-import com.createver.server.domain.image.service.ImageAvatarSseService;
+import com.createver.server.domain.image.service.avatar.ImageAvatarProcessingService;
+import com.createver.server.domain.image.service.avatar.ImageAvatarService;
+import com.createver.server.domain.image.service.avatar.ImageAvatarSseService;
 import com.createver.server.global.util.aws.service.S3DownloadService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
@@ -21,7 +21,6 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
 import static org.hamcrest.Matchers.containsString;
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
