+++ src/test/java/com/createver/server/domain/image/controller/ImageTagControllerTest.java
@@ -1,7 +1,7 @@
 package com.createver.server.domain.image.controller;
 
 import com.createver.server.domain.image.dto.ImageTagDto;
-import com.createver.server.domain.image.service.ImageTagService;
+import com.createver.server.domain.image.service.tag.ImageTagService;
 import com.createver.server.global.config.SecurityConfig;
 import com.createver.server.global.user.WithMockCustomMember;
 import org.junit.jupiter.api.DisplayName;
