+++ src/test/java/com/createver/server/domain/image/service/ImageAvatarSseServiceTest.java
@@ -1,5 +1,6 @@
 package com.createver.server.domain.image.service;
 
+import com.createver.server.domain.image.service.avatar.ImageAvatarSseService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
