+++ src/test/java/com/createver/server/domain/image/service/ImageReportServiceTest.java
@@ -4,6 +4,7 @@ import com.createver.server.domain.image.entity.Gallery;
 import com.createver.server.domain.image.entity.ImageReport;
 import com.createver.server.domain.image.repository.gallery.GalleryRepository;
 import com.createver.server.domain.image.repository.report.ImageReportRepository;
+import com.createver.server.domain.image.service.report.ImageReportService;
 import com.createver.server.domain.member.entity.Member;
 import com.createver.server.domain.member.repository.MemberRepository;
 import org.junit.jupiter.api.BeforeEach;
